google oauth done

+++ controllers/api/v1/posts_api.js
@@ -17,22 +17,25 @@ module.exports.index = async function(req, res){
         posts: posts
     })
 }
-
-
 module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
+        if (post.user == req.user.id){
             post.remove();
-
             await Comment.deleteMany({post: req.params.id});
-            return res.json(200 ,{
-                message:"Post deleted"
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!"
             });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
     }catch(err){
+        console.log('********', err);
         return res.json(500, {
-            message:"Internal Server Error"
-        })
-    }
-    
+            message: "Internal Server Error"
+        });
+    } 
 }
